Added code cov and tests for storage
